Split tests into extrinsic, helper and utils.
